+++ cmd/river/rivercli/river_cli_test.go
@@ -1,4 +1,4 @@
-package main
+package rivercli
 
 import (
 	"testing"
